// src/components/Menu/UikitMenu/types.ts
import React, { ReactElement } from 'react'
import { Colors, SvgProps } from '@pancakeswap/uikit'

export interface Language {
  code: string
  language: string
  locale: string
}

export interface PushedProps {
  isPushed: boolean
  pushNav: (isPushed: boolean) => void
}

export interface NavTheme {
  background: string
}

export interface LinkStatus {
  text: string
  color: keyof Colors
}

export interface MenuSubEntry {
  label: string
  href: string
  calloutClass?: string
  status?: LinkStatus
}

export interface MenuEntry {
  label: string
  icon: string | React.FC<SvgProps>
  activeIcon?: string | React.FC<SvgProps>
  items?: MenuSubEntry[]
  href?: string
  calloutClass?: string
  initialOpenState?: boolean
  status?: LinkStatus
}

export interface PanelProps {
  isDark: boolean
  toggleTheme: (isDark: boolean) => void
  cakePriceUsd?: number
  currentLang: string
  langs: Language[]
  setLang: (lang: Language) => void
  links: Array<MenuEntry>
}

export interface NavProps extends PanelProps {
  userMenu?: ReactElement
  globalMenu?: ReactElement
}